chore: funcao stack_copy ex8
