+++ Algoritmos e Estrutura de Dados/p1/stack/stack.h
@@ -16,5 +16,6 @@ bool stack_isEmpty(Stack* stack);
 bool stack_is_full(Stack* stack);
 int stack_size(Stack* stack);
 void stack_print(Stack* stack);
+Stack* stack_copy(Stack* s);
 
 #endif
